ath10k: Tweak flushing logic for CT firmware.
Don't bother waiting if we are using CT firmware.  CT
firmware can force a flush (and it will just drop them),
so don't bother waiting for the entire firmware tx-queue to
empty.

Likely that will not work anyway when there are other vdevs
around, so waiting just wastes 3 seconds under a lock for
no good reason.

Signed-off-by: Ben Greear <greearb@candelatech.com>