ath10k:  Support fwtest API, allow setting sounding timers.
Upstream firmware has the ability to set the MU-MIMO and SU-MIMO
sounding timers, through two different API.  The MU-MIMO timer
is set through the 'fwtest' API, which is basically a random
collection of hacks with hard-coded IDs (no enums in sight).

Implement the fwtest API for 10.4 firmware, and add hooks to the
ct_special API to allow setting su-mimo and mu-mimo.

Signed-off-by: Ben Greear <greearb@candelatech.com>